[supervisor] Move out of the Theia folder
