+++ components/supervisor-api/generate.sh
