+++ gitpod-ws.theia-workspace
@@ -4,15 +4,15 @@
       { "path": "components/common-go" },
       { "path": "components/content-service" },
       { "path": "components/ee/cerc" },
-      { "path": "components/ws-manager" },
       { "path": "components/ee/ws-scheduler" },
       { "path": "components/gitpod-cli" },
       { "path": "components/image-builder" },
       { "path": "components/licensor" },
       { "path": "components/registry-facade" },
       { "path": "components/service-waiter" },
-      { "path": "components/theia/supervisor" },
+      { "path": "components/supervisor" },
       { "path": "components/ws-manager-node" },
+      { "path": "components/ws-manager" },
       { "path": "components/ws-proxy" },
       { "path": "components/ws-sync" },
       { "path": "install/installer" }
